Consolidated the exchange tests into a single file again.
